Turn module into a package.
This way we get a directory (the package dir) which we can use to
store data files, which otherwise are difficult to find after
installation. Data files would be installed in different locations by
different tools (setuptools, pip).

To avoid all that installatin trouble we simply put all files into
the package and declare them as package data.